move kubeconfig generator into `hooks/eks.py`
